pyproject.toml
[project]
name = "lss-daily-discord-overview"
version = "2.2.0"
description = "Dieses Skript ruft Daten von Leitstellenspiel.de ab und sendet eine tägliche Übersicht zu Gebäude-Erweiterungen und Schulungen, die heute fertig werden, an einen Discord-Webhook."
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E"]
authors = [
    { name = "Mister_X", email = "MisterX2000@users.noreply.github.com" },
    { name = "L0rdEnki", email = "77723580+L0rdEnki@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "apprise~=1.9.5",
    "beautifulsoup4~=4.14.3",
    "inquirer~=3.4.1",
    "python-dotenv~=1.2.1",
    "requests~=2.32.5",
]

[dependency-groups]
dev = [
    "pyinstaller~=6.17.0",
]

[project.scripts]
lss-daily-discord-overview = "lss_daily_discord_overview:main"

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[tool.bumpversion]
current_version = "2.2.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = true
tag_name = "v{new_version}"
allow_dirty = false
commit = true
message = "Bump version: {current_version} → {new_version}"
pre_commit_hooks = ["uv sync --upgrade", "git add uv.lock"]
